nieznaczne zmiany stuktury w zwiazku ze wsteczna kompatybilnoscia

+++ database/migrations/2016_01_18_191946_create_polls_table.php
@@ -17,7 +17,7 @@ class CreatePollsTable extends Migration
             $table->string('title');
             $table->timestampTz('created_at')->default(DB::raw('CURRENT_TIMESTAMP(0)'));
             $table->timestampTz('updated_at')->nullable();
-            $table->smallInteger('length')->default(0);
+            $table->integer('length')->default(0);
             $table->smallInteger('max_items')->default(1);
         });
